==> 1_Parsing_and_Mering_AllData.ipynb <==
## Merge Raw Petfinder Data
207 Final Project<br>
Lucy Herr<br>
7/13/23<br>
### Process Overview
1. Parse Pet Listing Image Metadata
2. Parse Pet Listing Sentiment Analysis Data
3. Merge Image & Sentiment Data with Tabular Data
4. Add external data features (State Population & Median Income)

### Import Packages

# import packages
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import csv
import json
import flatten_json
import os
import warnings
import wget
import urllib3
from sklearn.model_selection import train_test_split # to split the data

# import cleaning_functions
import cleaning_functions # assumes cleaning functions file is in same working directory as this notebook

### Load data & functions

# data directory path 
data_dir = './data/' # assumes data subdirectory (follows structure of original file)

# load raw tabular training data 
train_file = 'train/train.csv' # assumes data & train subdirectories (follows structure of original file)
train_tab_df = pd.read_csv(data_dir+train_file)

# load raw image metadata (JSON)
image_metadata_path = 'train_metadata/' 

# load raw sentiment metadata (JSON)
sentiment_path = 'train_sentiment/'
sentiment_json_files = [pos_json for pos_json in os.listdir(data_dir+sentiment_path) 
                        if pos_json.endswith('.json')]

# Malaysian state labels 
state_file = 'state_labels.csv'
state_labels = pd.read_csv(data_dir+state_file)

# Malaysia population by state (external data source)
state_pop_url="https://raw.githubusercontent.com/MoH-Malaysia/covid19-public/main/static/population.csv"
state_pop_df=pd.read_csv(state_pop_url)

# Malaysia median income by state (external data source)
income_url = 'https://storage.googleapis.com/dosm-public-economy/salaries_state_sex.csv'
state_income_df=pd.read_csv(income_url)


income_url = 'https://storage.googleapis.com/dosm-public-economy/salaries_state_sex.csv'
state_income_df=pd.read_csv(income_url)
state_income_df.head()

print(train_tab_df.shape)
train_tab_df.columns

### 1. Parse Pet Listing Image Metadata & Add to Tabular Dataframe

# set path to image metadata directory
image_metadata_path = 'train_metadata/' 

# lists to store each image metadata feature prior to EDA
vertex_xs = []
vertex_ys = []
bounding_confidences = []
bounding_importance_fracs = []
dominant_blues = []
dominant_greens = []
dominant_reds = []
dominant_pixel_fracs = []
dominant_scores = []
label_descriptions = []
label_scores = []
face_annotations = []
label_annotations = []
text_annotations = []

# lists of pet IDs without label annotations and/or any image metadata
petids_no_image_metadata = []
petids_no_label_annotation = []

# store each feature value for each pet ID in the main (tabular) df 
for pet_id in train_tab_df.PetID:
    try:
        # NOTE: selecting 1st image (when multiple images are available)
        with open(data_dir+image_metadata_path + pet_id + '-1.json', 'r') as f: 
            data = json.load(f)
        face_annotations.append(data.get('faceAnnotations', []))
        text_annotations.append(data.get('textAnnotations', []))
        vertex_x = data['cropHintsAnnotation']['cropHints'][0]['boundingPoly']['vertices'][2]['x']
        vertex_xs.append(vertex_x)
        vertex_y = data['cropHintsAnnotation']['cropHints'][0]['boundingPoly']['vertices'][2]['y']
        vertex_ys.append(vertex_y)
        bounding_confidence = data['cropHintsAnnotation']['cropHints'][0]['confidence']
        bounding_confidences.append(bounding_confidence)
        bounding_importance_frac = data['cropHintsAnnotation']['cropHints'][0].get('importanceFraction', -1)
        bounding_importance_fracs.append(bounding_importance_frac)
        dominant_blue = data['imagePropertiesAnnotation']['dominantColors']['colors'][0]['color']['blue']
        dominant_blues.append(dominant_blue)
        dominant_green = data['imagePropertiesAnnotation']['dominantColors']['colors'][0]['color']['green']
        dominant_greens.append(dominant_green)
        dominant_red = data['imagePropertiesAnnotation']['dominantColors']['colors'][0]['color']['red']
        dominant_reds.append(dominant_red)
        dominant_pixel_frac = data['imagePropertiesAnnotation']['dominantColors']['colors'][0]['pixelFraction']
        dominant_pixel_fracs.append(dominant_pixel_frac)
        dominant_score = data['imagePropertiesAnnotation']['dominantColors']['colors'][0]['score']
        dominant_scores.append(dominant_score)
        # add label annotation data 
        if data.get('labelAnnotations'):
            label_annotations.append(data['labelAnnotations'])
            label_description = data['labelAnnotations'][0]['description']
            label_descriptions.append(label_description)
            label_score = data['labelAnnotations'][0]['score']
            label_scores.append(label_score)
        else:
            # append pet id to missing annotation list
            petids_no_label_annotation.append(pet_id)
            label_annotations.append(None)
            label_descriptions.append(None) # blank field = none
            label_scores.append(-1)
    except FileNotFoundError:
        # append pet id to missing annotation list
        petids_no_image_metadata.append(pet_id)
        vertex_xs.append(-1)
        vertex_ys.append(-1)
        bounding_confidences.append(-1)
        bounding_importance_fracs.append(-1)
        dominant_blues.append(-1)
        dominant_greens.append(-1)
        dominant_reds.append(-1)
        dominant_pixel_fracs.append(-1)
        dominant_scores.append(-1)
        label_annotations.append([None])
        label_descriptions.append(None) # blank field = none
        label_scores.append(-1)
        face_annotations.append([None])
        text_annotations.append([None])

print(f'Number of PetIDs missing label annotations: {len(petids_no_label_annotation)}')
print(f'Number of PetIDs missing image metadata: {len(petids_no_image_metadata)}')

# add new feature columns to main (tabular) dataframe
# dictionary of the new columns and lists to concat
image_data = {
    'vertex_xs': vertex_xs,
    'vertex_ys': vertex_ys,
     'bounding_confidences': bounding_confidences,
     'bounding_importance_fracs': bounding_importance_fracs,
     'dominant_blues': dominant_blues,
     'dominant_greens': dominant_greens,
     'dominant_reds': dominant_reds,
     'dominant_pixel_fracs': dominant_pixel_fracs,
     'dominant_scores': dominant_scores,
     'label_descriptions': label_descriptions,
     'label_scores': label_scores
}

#add new feature columns to main (tabular) dataframe
train_tab_df = pd.concat([
    train_tab_df,
    pd.DataFrame(image_data)],
    axis=1)
print(train_tab_df.shape)
print(train_tab_df.shape)

# for f in image_data:
#     print(f"{f} has length {len(image_data[f])}")
train_tab_df.shape[0] -341


### 2. Parse Pet Listing Sentiment Analysis Data & Add to Dataframe

data = []

sentiment_path = './data/train_sentiment' 
json_files = [pos_json for pos_json in os.listdir(sentiment_path) if pos_json.endswith('.json')]

for js_file in json_files:
    with open(os.path.join(sentiment_path, js_file)) as f:
        data.append(json.load(f))

id_list = [f.split('.')[0] for f in json_files]  # assuming file name is petID
sent_magnitudes = [[sentence['sentiment']['magnitude'] for sentence in d['sentences']] if 'sentences' in d else None for d in data]
sent_scores = [[sentence['sentiment']['score'] for sentence in d['sentences']] if 'sentences' in d else None for d in data]
doc_magnitudes = [d['documentSentiment']['magnitude'] if 'documentSentiment' in d else None for d in data]
doc_scores = [d['documentSentiment']['score'] if 'documentSentiment' in d else None for d in data]
languages = [d['language'] if 'language' in d else None for d in data]
sent_texts = [[sentence['text']['content'] for sentence in d['sentences']] if 'sentences' in d else None for d in data]

sentiment_df = pd.DataFrame({
    'petID': id_list,
    'sent_scores': sent_scores,
    'sent_magnitudes': sent_magnitudes,
    'doc_scores': doc_scores,
    'doc_magnitudes': doc_magnitudes,
    'languages': languages,
    'text':sent_texts
})json_files

train_tab_df = pd.merge(train_tab_df, sentiment_df, how='left', left_on='PetID', right_on='petID')

print(train_tab_df.shape)
train_tab_df.head()

len(json_files)

### 3. Add external data features (Malaysia State Population & Median Income)

#### Review state labels for state IDs & state names in Petfinder tabular data

print(f"""Number of unique StateID values in train set:\n {len(train_tab_df.State.unique())}\n""")
print(f"""unique StateID values in train set:\n{train_tab_df.State.unique()}\n""")

#rename 'State' col to 'StateID' for joining
train_tab_df.rename(columns={'State':'StateID'},inplace=True)
# check nulls in tabular data
print(len(train_tab_df[train_tab_df['StateID'].isnull()]))

#### Review state names for State IDs in Petfinder tabular data

print(f"""Number of unique State Names in Petfinder state labels:\n{len(state_labels.StateName.value_counts())}\n""")
print(f"""Unique State Names in Petfinder state labels:\n{state_labels.StateName.value_counts()}""")
# state_labels = state_labels.sort_values(by='StateName')
# state_labels.reset_index(inplace=True)

#### Review state income data (median income by state, 2020)

state_income_df.head()

# use 'variable_en' as state name
state_income_df.rename(columns={'variable_en':'state'},inplace=True)
state_income_df = state_income_df.sort_values(by='state')
print(f'Number of unique state names in state income data: {len(state_income_df.state.value_counts())}')
print(state_income_df.state.value_counts())

# subset 2020 data
state_income_df = state_income_df.loc[state_income_df['year']==2020]
# subset overall data (all genders)
state_income_df = state_income_df.loc[state_income_df['sex']=='overall'] 
# remove 'overall' (total) row
state_income_df = state_income_df.loc[state_income_df['state']!='Overall']

# remove 'W.P.' prefix to match state names in petfinder labels 
state_income_df['state'] = state_income_df['state'].str.replace('W.P. ','')
                                                          
state_income_df = state_income_df.loc[:,['state','median']]

# sort by English state names 
state_income_df = state_income_df.sort_values(by='state')

state_income_df.reset_index(drop=True,inplace=True)

print(f'Number of unique state names in clean state income data: {len(state_income_df.state.value_counts())}')
print(state_income_df.state.value_counts())
# 'Putrajaya' not present in other state dfs


#### Review state population data (total population by state, 2020)

state_pop_df.head()

print(f'Number of unique state names in state pop. data: {len(state_pop_df.state.value_counts())}')
print(state_pop_df.state.value_counts())

# remove 'Malaysia' (total) row
state_pop_df = state_pop_df[state_pop_df['state'] != 'Malaysia'] 

# remove 'W.P.' prefix
state_pop_df['state'] = state_pop_df['state'].str.replace('W.P. ','')

state_pop_df = state_pop_df.sort_values(by='state')

print(f'Number of unique state names in cleaned state pop. data: {len(state_pop_df.state.value_counts())}')
print(state_pop_df.state.value_counts())
# 'Klang Valley' not present in other state df columns 

# check difference between population df & income df states
print("state names in income df but not in population df:")
print (list(set(state_income_df.state.unique()) - set(state_pop_df.state.unique())))

# check difference between income df and population df: 
print("state names in population df but not in income df:")
print (list(set(state_pop_df.state.unique())- set(state_income_df.state.unique())))

# remove trailing space after 'Pulau Pinang ' &  'Pahang ' in income df
state_income_df['state'] = state_income_df['state'].str.replace('Pulau Pinang ','Pulau Pinang')
state_income_df['state'] = state_income_df['state'].str.replace('Pahang ','Pahang')

# check that this worked
# check difference between population df & income df states
print("state names in income df but not in population df:")
print (list(set(state_income_df.state.unique()) - set(state_pop_df.state.unique())))

print("state names in population df but not in income df:")
print (list(set(state_pop_df.state.unique())- set(state_income_df.state.unique())))

state_pop_df.reset_index(drop=True,inplace=True)
# merge external state data
comb_state_df = pd.merge(state_pop_df,state_income_df,how='left',on='state')
# select relevant columns 
comb_state_df = comb_state_df.loc[:,['state','pop','median']]

print(f'Number of unique state names in combined state data: {len(comb_state_df.state.value_counts())}')
print(f'Number of unique state names in P.F. State Labels: {len(state_labels.StateName.value_counts())}')

# check discrepancies in state names between combined income/pop data and PF labels
print("State names in median/income df but not in Petfinder labels:")
print (list(set(comb_state_df.state.unique())- set(state_labels.StateName.unique())))

print("State names in Petfinder labels but not in median/income df:")
print (list(set(state_labels.StateName.unique()) - set(comb_state_df.state.unique())))

# rename columns for clarity 
comb_state_df.rename(columns={"state":"StateName",
                              "pop":"state_population",
                               "median":"median_state_income"}, 
                      inplace=True)

# merge with petfinder state labels 
state_labels = pd.merge(state_labels,comb_state_df,on='StateName')
state_labels

#### Join petfinder state labels with combined external state data

# join petfinder state labels with external state data 
train_tab_df = pd.merge(train_tab_df,state_labels,how='left', on='StateID')
#train_tab_df= train_tab_df.set_index('StateID').join(state_labels.set_index('StateID'))
print(train_tab_df.shape)
print(train_tab_df.columns)

train_tab_df.median_state_income.value_counts()

### -------begin Bailey added code section-------

# Split the combined data into test and train (80/20)

#### Split data into test/train

train_tab_df.AdoptionSpeed.value_counts(normalize=True)

# split the data into test and train (split training 80% so that 25% of it can be used for validation)

# Original version:
train, test = train_test_split(train_tab_df, test_size=0.2, random_state=42)

# Stratified version:
#train, test = train_test_split(train_tab_df, test_size=0.2,random_state=42, 
#                               stratify = train_tab_df.AdoptionSpeed) # stratify to balance labels in train/test sets


# review shape of data
print('Shape train: ', train.shape, '\nShape test: ', test.shape)
      #,'\nShape validation: ', validation.shape)

train.AdoptionSpeed.value_counts(normalize=True)

#### Write data to respective files
Note: The files will split the same way each time as long as you do not change the "random_state" in the train_test_split function.


# train
train.to_csv('split_train.csv')

# test 
test.to_csv('split_test.csv')
